Extract and clean up print view parser logic.
